вніс правки від ментора
