+++ server.js
@@ -5,11 +5,13 @@ const { DB_HOST, PORT = 3000 } = process.env;
 
 mongoose
   .connect(DB_HOST)
-  .then(console.log("Database connection successful"))
   .then(() =>
-    app.listen(PORT, () => console.log(`server is listening on port: ${PORT}`))
+    app.listen(PORT, () => {
+      console.log(`Database connection successful`);
+      console.log(`server is listening on port: ${PORT}`)
+    })
   )
   .catch((err) => {
     console.log("Error:", err.message);
     process.exit(1);
-  });
+  });
